Ascan API - Return alert count for each scanner
Change ActiveScanAPI to return the count of alerts raised by each
scanner.
Change HostProcess to expose the PluginStats (allow consumers to obtain
all the data, instead of adding more methods to HostProcess).
Move PluginStats to a top level class.
